+++ bin/commands/diff.js
@@ -1,30 +1,26 @@
-import fs from 'fs'
-import _ from 'lodash'
+import * as git from '../../src/git.js'
+
 import Codeowners from '../../src/codeowners.js'
+import _ from 'lodash'
+import { allowMultipleValues } from './run.js'
+import { countByMetric } from '../helpers.js'
+import { findOccurrences } from '../../src/occurrences.js'
+import fs from 'fs'
 import { getConfiguration } from '../../src/configuration.js'
 import { getFiles } from '../../src/files.js'
-import { findOccurrences } from '../../src/occurrences.js'
-import { countByMetric } from '../helpers.js'
-import * as git from '../../src/git.js'
 import { panic } from '../../src/error.js'
 
 export default function (program) {
   program
     .command('diff')
-    .requiredOption('--metric <metric>', 'Add a metric', (value, previous) =>
-      previous ? [...previous, value] : [value]
-    )
+    .requiredOption('--metric <metric>', 'will only consider provided metrics', allowMultipleValues)
     .option('--input-file <input_file>', 'A JSON file containing the metrics to compare with')
-    .option('--api-key <api_key>', 'THIS OPTION IS DEPRECATED, DO NOT USE IT')
     .option('--error-if-increase', 'Return an error status code (1) if the metric increased since its last report')
     .option('--quiet', 'reduce output to a minimum')
     .action(async (options) => {
       const configuration = await getConfiguration()
-      const metrics = options.metric
+      const metricNames = options.metric
       const inputFile = options.inputFile
-
-      // TODO: Remove this when the --api-key option is removed
-      if (options.apiKey) console.log('WARNING: --api-key is deprecated and will raise an error in the future.')
 
       let lastMetricValue
       let previousOccurrences
@@ -40,40 +36,44 @@ export default function (program) {
       const currentOccurrences = await findOccurrences({
         configuration,
         files: await getFiles(),
+        metricNames,
         codeOwners: new Codeowners(),
         quiet: options.quiet,
       })
 
       // TODO: If a file has been provided, then we can skip the merge base logic
       if (!inputFile) {
+        // Checkout the base branch to calculate its occurrences
         const defaultBranchName = await git.getDefaultBranchName()
         const baseBranchCommit = await git.getMergeBase(initialBranch, defaultBranchName)
         await git.checkout(baseBranchCommit)
         previousOccurrences = await findOccurrences({
           configuration,
           files: await getFiles(),
+          metricNames,
           codeOwners: new Codeowners(),
           quiet: options.quiet,
         })
-        await git.checkout(initialBranch) // Bring user back to initial branch
+        // Bring user back to initial branch
+        await git.checkout(initialBranch)
       }
 
       // For each metric, compare the current occurrences with the previous ones
-      for (const metric of metrics) {
+      for (const metricName of metricNames) {
         try {
           console.log('-----------------------------------')
-          console.log(`Metric: ${metric}`)
+          console.log(`Metric: ${metricName}`)
 
           if (inputFile) {
             const content = fs.readFileSync(inputFile, 'utf8')
             const metrics = JSON.parse(content)
-            metricOccurrences = metrics.find((m) => m.name === metric)?.occurrences || []
+            metricOccurrences = metrics.find((m) => m.name === metricName)?.occurrences || []
             previousOccurrences = metricOccurrences
             lastMetricValue = _.sumBy(metricOccurrences, (occurrence) =>
               _.isNumber(occurrence.value) ? occurrence.value : 1
             )
           } else {
-            lastMetricValue = countByMetric(previousOccurrences)[metric] || 0
+            lastMetricValue = countByMetric(previousOccurrences)[metricName] || 0
           }
 
           if (!Number.isInteger(lastMetricValue)) {
@@ -86,7 +86,7 @@ export default function (program) {
           process.exit(1)
         }
 
-        const currentMetricValue = countByMetric(currentOccurrences)[metric] || 0
+        const currentMetricValue = countByMetric(currentOccurrences)[metricName] || 0
         console.log(`Current value: ${currentMetricValue}`)
 
         const diff = currentMetricValue - lastMetricValue
@@ -95,7 +95,7 @@ export default function (program) {
         // Log added occurrences if any
         if (diff > 0) {
           console.log('Added occurrences:')
-          const currentMetricOccurrences = currentOccurrences.filter((o) => o.metricName === metric)
+          const currentMetricOccurrences = currentOccurrences.filter((o) => o.metricName === metricName)
           const currentMetricOccurrencesTexts = currentMetricOccurrences.map((o) => o.text)
           const previousOccurrencesTexts = previousOccurrences.map((occurrence) => occurrence.text)
           console.log(currentMetricOccurrencesTexts.filter((x) => !previousOccurrencesTexts.includes(x)))