+++ test/cherry_diff.test.js
@@ -1,56 +1,49 @@
+import { afterAll, beforeAll, describe, expect, it } from 'vitest'
+import { execAsync, expectError } from './helpers.js'
+
 import fs from 'fs'
-import { exec } from 'child_process'
+import path from 'path'
 
 const TEMPORARY_FILE_PATH = 'test_temporary_file'
 
-beforeEach(() => {
-  if (fs.existsSync(TEMPORARY_FILE_PATH)) {
-    fs.unlinkSync(TEMPORARY_FILE_PATH)
-  }
-})
+const CHERRY_BIN_PATH = 'node ./../../../bin/cherry.js'
+
+const originalCwd = process.cwd()
+const fakeProjectPath = path.join(originalCwd, 'test/fixtures/project-one')
 
 describe('cherry diff', () => {
-  test('averts that --api-key is deprecated', (done) => {
-    exec('node bin/cherry.js diff --quiet --metric TODO --api-key test', (error, stdout) => {
-      expect(error).toBe(null)
-      expect(stdout).toContain('WARNING: --api-key is deprecated and will raise an error in the future.')
-      expect(stdout).toContain('Metric: TODO')
-      done()
-    })
+  beforeAll(() => {
+    process.chdir(fakeProjectPath)
+    if (fs.existsSync(TEMPORARY_FILE_PATH)) fs.unlinkSync(TEMPORARY_FILE_PATH)
   })
 
-  test('should exit with an error if --metric is missing', (done) => {
-    exec('node bin/cherry.js diff --quiet', (error, _stdout, stderr) => {
-      expect(error.code).toBe(1)
-      expect(stderr).toContain(`required option '--metric <metric>' not specified`)
-      done()
-    })
+  afterAll(() => process.chdir(originalCwd))
+
+  it('uses the default configuration', async () => {
+    const { stdout } = await execAsync(`${CHERRY_BIN_PATH} diff --quiet --metric "[loc] JavaScript"`)
+    expect(stdout).toContain('No .cherry.js file found, using default configuration...')
   })
 
-  test('can take multiple metrics', (done) => {
-    exec('node bin/cherry.js diff --quiet --metric TODO --metric "[loc] JavaScript"', (error, stdout) => {
-      expect(error).toBe(null)
-      expect(stdout).toContain('Metric: TODO')
-      expect(stdout).toContain('Metric: [loc] JavaScript')
-      done()
-    })
+  it('should exit with an error if --metric is missing', async () => {
+    const error = await expectError(`${CHERRY_BIN_PATH} diff --quiet`)
+    expect(error.stderr).toContain(`required option '--metric <metric>' not specified`)
   })
 
-  test('requires to commit changes before running cherry diff', (done) => {
+  it('can take multiple metrics', async () => {
+    const { stdout } = await execAsync(`${CHERRY_BIN_PATH} diff --quiet --metric TODO --metric "[loc] JavaScript"`)
+    expect(stdout).toContain('Metric: TODO')
+    expect(stdout).toContain('Metric: [loc] JavaScript')
+  })
+
+  it('requires to commit changes before running cherry diff', async () => {
     fs.writeFileSync(TEMPORARY_FILE_PATH, 'unexpected content')
-
-    exec('node bin/cherry.js diff --quiet --metric TODO', (error, _stdout, stderr) => {
-      expect(error.code).toBe(1)
-      expect(stderr).toContain('Please commit your changes before running cherry diff.')
-      done()
-    })
+    const error = await expectError(`${CHERRY_BIN_PATH} diff --quiet --metric TODO`)
+    expect(error.stderr).toContain('Please commit your changes before running cherry diff')
   })
 
-  test('does not require to commit changes when --input-file is provided', (done) => {
-    exec('node bin/cherry.js diff --quiet --metric TODO --input-file test --api-key test', (error, _stdout, stderr) => {
-      expect(error.code).toBe(1)
-      expect(stderr).not.toContain('Please commit your changes before running cherry diff.')
-      done()
-    })
+  it('does not require to commit changes when --input-file is provided', async () => {
+    const error = await expectError(`${CHERRY_BIN_PATH} diff --quiet --metric TODO --input-file ${TEMPORARY_FILE_PATH}`)
+    expect(error.code).toBe(1)
+    expect(error.stderr).not.toContain('Please commit your changes before running cherry diff')
   })
 })