+++ test/cherry_run.test.js
@@ -1,18 +1,16 @@
-import { exec } from 'child_process'
+import { describe, expect, it } from 'vitest'
+
+import { execAsync } from './helpers'
 
 describe('cherry run', () => {
-  test('runs all metrics from config file', (done) => {
-    exec('node bin/cherry.js run', (_error, stdout) => {
-      expect(stdout).toContain('JS circular dependencies')
-      expect(stdout).toContain('TODO')
-      done()
-    })
+  it('runs all metrics from config file', async () => {
+    const { stdout } = await execAsync('node bin/cherry.js run')
+    expect(stdout).toContain('JS circular dependencies')
+    expect(stdout).toContain('TODO')
   })
 
-  test('runs only selected metric', (done) => {
-    exec('node bin/cherry.js run --metric TODO', (_error, stdout) => {
-      expect(stdout).toContain('Total occurrences:')
-      done()
-    })
+  it('runs only selected metric', async () => {
+    const { stdout } = await execAsync('node bin/cherry.js run --metric TODO')
+    expect(stdout).toContain('Total occurrences:')
   })
 })